fix: 메소드명 불일치 및 로직 오류 수정

+++ src/main/java/com/fairing/fairplay/qr/repository/QrCheckStatusCodeRepository.java
@@ -4,6 +4,6 @@ import com.fairing.fairplay.qr.entity.QrCheckStatusCode;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QrCheckStatusCodeRepository extends JpaRepository<QrCheckStatusCode, Long> {
+public interface QrCheckStatusCodeRepository extends JpaRepository<QrCheckStatusCode, Integer> {
   Optional<QrCheckStatusCode> findByCode(String qrCheckStatusCode);
 }
